Add setting validation and Azure prompts back to C# templates (#458)
We lost setting validation and some Azure-specific prompts when we switched to dynamic retrieval of .NET templates. Fortunately, we can pretty easily copy this information from JavaScript templates. We only do this if the template name and setting name match exactly, so it will default to the plain-old .NET experience when applicable.

I also leveraged the 'Documenation' information that is supplied with the .NET templates. This is the one case where these templates actually provide more information than the JavaScript templates.