Python Project Creating/Local Debugging (#525)
Create and locally debug Python functions projects.  Option to create is still hidden as Python is still in preview and we are working on the deployment scenario.

In order to debug, `python -m pip install --pre ptvsd` must be run within the project's virtual environment.  There is also a dependency on the experimental Python Extension, but they should release on Tuesday, 9/5/2018.