Contribute func-watch problemMatcher (#578)
I'm basically shifting the implementation of the problem matcher from each user's project to our extension. This means we can change/improve it in the future without having to modify the user's project.

Unfortunately, the func cli is very annoying to parse, so I only added support for one error (although it's somewhat common).