Fix template source logic for C# tests (#787)
So I realized that the C# tests weren't actually running tests based on the different template sources since they overwrite the nuget package in the same location. The key fix was to change the location from:
```
path.join('resources', 'dotnetTemplates')
```
to
```
path.join('resources', 'dotnetTemplates', ext.templateSource || '')
```

The rest of the changes are based on that and a little bit of refactor. There should be no functional changes.