Update "verifyDebugConfigIsValid" to work with new tasks (#958)
The breaking changes in the func cli were way back in June, but we still see updatedDebugConfig occasionally in telemetry. Enough that I wanted to update this based on recent task changes, but not enough to spend much time on this. I just leverage initProjectForVSCode instead of fixing every little change manually