Unblock users with project in a sub folder (#1021)
We've had quite a few users run into issues when their project is in a sub folder of the root. Unfortunately, our 0.14.0 release made this slightly worse because we're providing tasks in `FuncTaskProvider.ts` and we only provide them if we detect the root as a function project. I don't have time to completely fix the issue, but this work will unblock users and they can follow my steps here to manually configure their project: https://github.com/Microsoft/vscode-azurefunctions/issues/686#issuecomment-466252387