Add new project support for F# (#1042)
With this project, users can debug and deploy. Biggest piece missing is creating a trigger - ultimately we'll have to wait for https://github.com/Azure/azure-functions-templates/issues/627. Because of that, I'm not showing "F#" in the prompt to create a project. User has to set "azureFunctions.projectLanguage" in user settings to "F#" _before_ creating a project to access this feature.