No longer recommend preDeployTask (#1201)
We originally added this logic specifically for `func extensions install` because it was not an obvious task and because it was added after many users had already created projects without it. Now - that task isn't required. I don't think it's worth keeping the logic for the other tasks, since those are language-specific, pretty obvious, and the vast majority of projects have them by default anyways.