Ask if venv should be created when initiating an existing Python project
Users are currently asked if a virtual environment (venv) should be created when
creating a new Python project, this also triggers creation of tasks for manage
builds within the venv. However, when importing/initiating an existing
project (e.g. when git cloned), the user is not prompted in the same manner
if a venv is to be created. This makes it harder to import an existing project.

I have copied the logic from the create flow to the init flow to provide the
user with the same options.