Add .Net 6 nightly test
The changes to "SubscriptionTreeItem" are because .NET 6 is the only test using FuncVersion v4, but when we do `getWorkspaceSettingFromAnyFolder`, it's returning v3 from a different folder. In the case of deploy, we know which folder is being used so we can be more specific than just "AnyFolder".

The changes to "test.yml" are to make sure the right .NET sdks are installed on the build machine