Use Uri.fspath as key for FuncRunningMap/Port
This commit modifies pickFuncProcess startFuncProcessFromApi removing
the workspaceFolder requirement and adding in the environment variables.

funcHostTask was originally tracked by vscode.WorkspaceFolder via
Task.Scope and DebugConfiguration.WorkspaceFolder. This commit modifies
it to use a new AzureFunctionTaskDefinition.functionsApp property which
is either the `buildPath` from startFuncProcessFromApi or the
vscode.WorkspaceFolder.uri.fsPath to maintain backwards compatibility.

runningFuncTaskMap, funcTaskStartedEmitter, and runningFuncPortMap now
all using the vscode.Uri.fsPath string which can be used to map back to
a WorkspaceFolder. However in the cases the WorkspaceFolder fails, APIs
that used it will now see undefined passed in.