Expose API command to start a func task and to check if CLI tools is installed (#4299)
* Add an API command to start a func task and to check if CLI tools is installed

* Make some updates for isFuncCoreToolsInstalled

* Update apiVersion

* Update apiVersion to minor path

* Use fully qualified path rather than relative

* Use Uri.fspath as key for FuncRunningMap/Port (#4309)

* Use Uri.fspath as key for FuncRunningMap/Port

This commit modifies pickFuncProcess startFuncProcessFromApi removing
the workspaceFolder requirement and adding in the environment variables.

funcHostTask was originally tracked by vscode.WorkspaceFolder via
Task.Scope and DebugConfiguration.WorkspaceFolder. This commit modifies
it to use a new AzureFunctionTaskDefinition.functionsApp property which
is either the `buildPath` from startFuncProcessFromApi or the
vscode.WorkspaceFolder.uri.fsPath to maintain backwards compatibility.

runningFuncTaskMap, funcTaskStartedEmitter, and runningFuncPortMap now
all using the vscode.Uri.fsPath string which can be used to map back to
a WorkspaceFolder. However in the cases the WorkspaceFolder fails, APIs
that used it will now see undefined passed in.

* Make unique 'type' so VS Code will launch multiple tasks

* Address Comments

* Pin core tools to fix tests (#4310)

* Try stable v4 feed

* Fixup

---------

Co-authored-by: Andrew Wang <waan@microsoft.com>
Co-authored-by: Alex Weininger <alex.weininger@live.com>